sentry - Application struct
